backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
websockets